corregidos los metodos CRUD para serverles de la madre

+++ api/routes/category.router.js
@@ -1,9 +1,8 @@
+require('dotenv').config();
 const express = require('express');
 const {auth, requiredScopes} = require('express-oauth2-jwt-bearer')
-
 const CategoryService = require('./../services/category.service');
-const validatorHandler = require('./../middlewares/validator.handler');
-const { createCategorySchema  } = require('./../schemas/category.schema');
+
 
 const router = express.Router();
 const service = new CategoryService();
@@ -15,62 +14,40 @@ const jwtCheck = auth({
 
 const checkScopes = requiredScopes(['read:endpoints']);
 
-router.get('/',jwtCheck,checkScopes, async (req, res, next) => {
+router.use(jwtCheck,checkScopes);
+
+router.all('/:id?',async(req,res,next)=>{
   try {
-    const categories = await service.find();
-    res.json(categories);
+    const { id } = req.params;
+    const body = req.body;
+
+    switch (req.method) {
+      case 'GET':
+        const category = id ? await service.findOne(id) : await service.find();
+        return res.json(category);
+      
+      case 'POST':
+        const newCategory = await service.create(body);
+        return res.status(201).json(newCategory);
+      
+      case 'PATCH':
+        if (!id) throw new Error("ID is required for updating a Category");
+        const updatedCategory = await service.update(id, body);
+        return res.json(updatedCategory);
+      
+      case 'DELETE':
+        if (!id) throw new Error("ID is required for deleting a Category");
+        const deletedCategory = await service.delete(id);
+        return res.json(deletedCategory);
+      
+      default:
+        res.status(405).json({ message: "Method Not Allowed" });
+    }
   } catch (error) {
     next(error);
   }
 });
 
-router.get('/:id',jwtCheck,checkScopes,
-  async (req, res, next) => {
-    try {
-      const { id } = req.params;
-      const category = await service.findOne(id);
-      res.json(category);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
 
-router.post('/',jwtCheck,checkScopes,
-  async (req, res, next) => {
-    try {
-      const body = req.body;
-      const newCategory = await service.create(body);
-      res.status(201).json(newCategory);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
-
-router.patch('/:id',jwtCheck,checkScopes,
-  async (req, res, next) => {
-    try {
-      const { id } = req.params;
-      const body = req.body;
-      const category = await service.update(id, body);
-      res.json(category);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
-
-router.delete('/:id',jwtCheck,checkScopes,
-  async (req, res, next) => {
-    try {
-      const { id } = req.params;
-      await service.delete(id);
-      res.status(201).json({id});
-    } catch (error) {
-      next(error);
-    }
-  }
-);
 
 module.exports = router;
